Separate some stuff into Concerns to keep the base generators not so huge
